Add a NixOS configuration for a rescue environment
This is done outside of the clan logic, and can be built with
`build-live-cd`. I am using that with Ventoy to setup a thumbdrive with
everything I would need to access my infrastructure, password vault, and
setup a computer from scratch in case something goes wrong with the
current one(s).