Add initial config for nsrv-cfr-scylla
With ZFS to hold data, still using some old school LVM for the system
disk, I did not have time to look into using ZFS on the root.

ZFS still seems ahead of the rest, in particular I'm doing some raidz2
with a hot spare, which is easier to manage than some RAID10 situation
on btrfs or mdadm.